pyproject.toml
[project]
name = "bot-detection-competition"
version = "0.1.0"
description = "Bot detection competition project for distinguishing AI-generated from human-written essays"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "scikit-learn>=1.3.0",
    "plotly>=5.15.0",
    "wordcloud>=1.9.0",
    "nltk>=3.8.0",
    "textstat>=0.7.0",
    "prefect>=2.14.0",
    "google-cloud-storage>=2.10.0",
    "kaggle>=1.5.16",
    "pyarrow>=14.0.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot_detection"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
strict = true
